refactor(identity): move Portal identity to portal/identity with one canonical resolution path
Closes #415. Refs #400, #403, #407.

- Move internal/identity to portal/identity; the SDK and every caller
  consume the owner directly with no forwarding wrappers.
- Collapse store.go into one canonical path: Resolve normalizes and
  derives key material, verifies supplied address/public key once, and
  fills the token secret; Generate/Parse/Marshal build on it.
- Delete ResolveListenerIdentity, loadOrCreateIdentity, resolveExposeName,
  and the save -> reload -> revalidate cycles; store.go is gone.
- Own CLI identity policy (flag precedence, name defaulting, create-on-
  absent) in cmd/portal-tunnel/identityfile, composed by the CLI, agent,
  demo-app, and payment-app.
- Keep relay persistence relay-owned and simple: LoadOrCreateRelayIdentity
  loads or generates, normalizes once in memory, and writes once.
- sdk.Expose requires an already resolved identity and validates through
  identity.Resolve; ExposeConfig loses IdentityPath/IdentityJSON.
- Add canonical-path, relay-identity, and CLI-policy tests.
