+++ cmd/payment-app/main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
-	"github.com/gosuda/portal-tunnel/v2/internal/identity"
+	"github.com/gosuda/portal-tunnel/v2/cmd/portal-tunnel/identityfile"
 	"github.com/gosuda/portal-tunnel/v2/sdk"
 	"github.com/gosuda/portal-tunnel/v2/types"
 	"github.com/gosuda/portal-tunnel/v2/utils"
@@ -151,20 +151,9 @@ func runPaymentApp(ctx context.Context, cfg paymentConfig) error {
 		return err
 	}
 
-	listenerIdentity, createdIdentity, err := identity.ResolveListenerIdentity(
-		types.Identity{Name: cfg.name},
-		cfg.addr,
-		cfg.identityPath,
-		cfg.identityJSON,
-	)
+	listenerIdentity, err := identityfile.Resolve(cfg.name, cfg.addr, cfg.identityPath, cfg.identityJSON)
 	if err != nil {
 		return fmt.Errorf("resolve identity: %w", err)
-	}
-	if createdIdentity {
-		log.Info().
-			Str("identity_path", strings.TrimSpace(cfg.identityPath)).
-			Str("address", listenerIdentity.Address).
-			Msg("generated tunnel identity and saved it to disk")
 	}
 	exposure, err := sdk.Expose(ctx, sdk.ExposeConfig{
 		RelayURLs:       utils.SplitCSV(cfg.relayURLs),
