+++ portal/acme/acme.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/go-acme/lego/v4/certificate"
 
-	"github.com/gosuda/portal-tunnel/v2/internal/identity"
+	"github.com/gosuda/portal-tunnel/v2/portal/identity"
 	"github.com/gosuda/portal-tunnel/v2/types"
 	"github.com/gosuda/portal-tunnel/v2/utils"
 )
