+++ portal/overlay/runtime_test.go
@@ -18,7 +18,7 @@ import (
 	"gosuda.org/ivnp/foundation"
 
 	"github.com/gosuda/portal-tunnel/v2/internal/discovery"
-	"github.com/gosuda/portal-tunnel/v2/internal/identity"
+	"github.com/gosuda/portal-tunnel/v2/portal/identity"
 	"github.com/gosuda/portal-tunnel/v2/portal/policy"
 	"github.com/gosuda/portal-tunnel/v2/types"
 )
