+++ sdk/expose.go
@@ -14,7 +14,7 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/gosuda/portal-tunnel/v2/internal/discovery"
-	"github.com/gosuda/portal-tunnel/v2/internal/identity"
+	"github.com/gosuda/portal-tunnel/v2/portal/identity"
 	"github.com/gosuda/portal-tunnel/v2/types"
 	"github.com/gosuda/portal-tunnel/v2/utils"
 )
@@ -125,7 +125,7 @@ func Expose(ctx context.Context, cfg ExposeConfig) (*Exposure, error) {
 		strings.TrimSpace(cfg.Identity.PrivateKey) == "" {
 		return nil, errors.New("portal sdk: identity must include address, public key, and private key")
 	}
-	listenerIdentity, _, err := identity.ResolveListenerIdentity(cfg.Identity.Copy(), "", "", "")
+	listenerIdentity, err := identity.Resolve(cfg.Identity.Copy())
 	if err != nil {
 		return nil, fmt.Errorf("resolve identity: %w", err)
 	}
